consolidate useTreeStore (hook) and useDecisionTree hook (we still have the useTreeStore hook that connects to our global store)
