+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export { useDecisionTree } from './useDecisionTree/useDecisionTree';
-export { useTreeStore } from 'hooks/useTreeStore/useTreeStore';
 export { useFetchConfig } from './useFetchConfig/useFetchConfig';
 export { useTreeDirection } from './useTreeDirection/useTreeDirection';
 export { useTreeViewport } from './useTreeViewport/useTreeViewport';
